rag-system/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
pypdf>=3.17.4
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.21
rank-bm25>=0.2.2
numpy>=1.26.0
aiofiles>=23.2.1
httpx>=0.26.0
ollama>=0.2.0
# Database
sqlalchemy>=2.0.0
# OCR dependencies (optional - only needed if using OCR)
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
